+++ svr/svr-ehr-basic/src/main/java/com.yihu.ehr/tj/dao/XTjDimensionSlaveRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
  * @version 1.0
  * @created 2017/5/08
  */
-public interface XTjDimensionSlaveRepository extends PagingAndSortingRepository<TjDimensionSlave, Integer> {
+public interface XTjDimensionSlaveRepository extends PagingAndSortingRepository<TjDimensionSlave, Long> {
 
 }
